web/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
jinja2>=3.1
google-cloud-firestore>=2.16
google-cloud-storage>=2.14
pydantic>=2.5
pydantic-settings>=2.1
